+++ lib/ipfs.ts
@@ -53,3 +53,8 @@ export const store = async (
   }
   return uploadToIpfs(createJsonBlob(nft));
 };
+
+export const getIpfsLink = (hash: string) =>
+  hash?.indexOf?.("ipfs://") > -1
+    ? hash.replace("ipfs://", "https://nftstorage.link/ipfs/")
+    : hash;